[SYCL][Reduction] Optimize reduCGFuncForRangeFastAtomics for discrete GPU
Main idea is that memory transfer host->GPU has significant cost here unlike the
integrated GPU/shared memory. As such, make sure sycl::no_init is used for
partial sums buffer (free change). Also, modify initialization of group counter
buffer to do it with an extra kernel as that is cheaper than transferring memory
host->device.

By itself the change would have caused incompatibilities in PartialSums's
accessor type. In order to deal with this, the optimization for a single
WorkGroup was moved from the host to device code. That also made possible to
always write final user's variable inside main kernel avoiding calling
reduSaveFinalResultToUserMem for this scenario. As such, changed
reduCGFuncForRange* to return a boolean indicating if such post-processing is
needed as that is a property of a particular implementation now.