[NFCI][SYCL] Use `public` inheritance from `[Uns|S]ampledImageAccessorBaseHost`
All the members are being moved to `protected` to account for that. This
is necessary for a subsequent change to implement `ObjBase` which needs
to have `public` inheritance (with members being hidden as `protected`
inside) so that `getSyclObjFromImpl` could access that base from a full
SYCL object.

Making this a separate PR so that the later one would be more mechanical.